Get rid of the task-reminder-selector view and include it into the overview.py
