+++ opengever/task/browser/overview.py
@@ -4,6 +4,8 @@ from opengever.base.browser.helper import get_css_class
 from opengever.globalindex.model.task import Task
 from opengever.tabbedview import GeverTabMixin
 from opengever.task import _
+from opengever.task.reminder import TASK_REMINDER_OPTIONS
+from opengever.task.reminder.reminder import TaskReminder
 from opengever.task.task import ITask
 from opengever.tasktemplates.interfaces import IFromParallelTasktemplate
 from opengever.tasktemplates.interfaces import IFromSequentialTasktemplate
@@ -12,10 +14,16 @@ from plone.app.contentlisting.interfaces import IContentListing
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser import BrowserView
+from zope.browserpage.viewpagetemplatefile import ViewPageTemplateFile
+from zope.i18n import translate
+import json
 
 
 class Overview(BrowserView, GeverTabMixin):
     """Provide / override tabbedview methods for task overviews."""
+
+    task_reminder_sleector_template = ViewPageTemplateFile(
+        'templates/task_reminder_selector.pt')
 
     def documents(self):
         """Return containing documents and related documents."""
@@ -107,7 +115,7 @@ class Overview(BrowserView, GeverTabMixin):
             {
                 'label': _(u"label_reminder", default=u"Reminder"),
                 'css_class': "taskReminderSelector",
-                'value': self.context.restrictedTraverse('@@task_reminder_selector')(),
+                'value': self.task_reminder_sleector_template(self),
                 'is_html': True,
                 },
             {
@@ -198,3 +206,43 @@ class Overview(BrowserView, GeverTabMixin):
 
     def previous_task_link(self):
         return self.render_task(self.context.get_sql_object().get_previous_task())
+
+    def current_reminder_title(self):
+        for option in self.reminder_options():
+            if option.get('selected'):
+                return option.get('option_title')
+
+    def reminder_options(self):
+        options = []
+        reminder_option = TaskReminder().get_reminder(self.context)
+
+        options.append({
+            'option_type': 'no-reminder',
+            'option_title': translate(_('no_reminder', default='No reminder'),
+                                      context=self.request),
+            'sort_order': -1,
+            'selected': reminder_option is None,
+            })
+
+        for option in TASK_REMINDER_OPTIONS.values():
+            selected = option.option_type == reminder_option.option_type if \
+                reminder_option else None
+            options.append({
+                'option_type': option.option_type,
+                'sort_order': option.sort_order,
+                'option_title': translate(
+                    option.option_title, context=self.request),
+                'selected': selected,
+            })
+
+        return options
+
+    def reminder_init_state(self):
+        return json.dumps({
+            'endpoint': self.context.absolute_url() + '/@reminder',
+            'reminder_options': self.reminder_options(),
+            'error_msg': translate(_('error_while_updating_task_reminder',
+                                     default="There was an error while "
+                                             "updating the reminder"),
+                                   context=self.request)
+        })
