+++ opengever/task/configure.zcml
@@ -17,7 +17,6 @@
   <include file="behaviors.zcml" />
   <include package=".upgrades" />
   <include package=".response_syncer" />
-  <include package=".reminder" />
 
   <vdex:vocabulary directory="vdexvocabs" />
