+++ opengever/task/locales/de/LC_MESSAGES/opengever.task.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: 1.0\n"
-"POT-Creation-Date: 2018-10-04 15:24+0000\n"
+"POT-Creation-Date: 2018-10-05 07:33+0000\n"
 "PO-Revision-Date: 2015-02-17 18:41+0100\n"
 "Last-Translator: Jonas Baumann <j.baumann@4teamwork.ch>\n"
 "Language-Team: 4teamwork <info@4teamwork.ch>\n"
@@ -227,7 +227,7 @@ msgid "error_same_responsible"
 msgstr "Keine Änderung, da kein anderer Auftragnehmer ausgewählt wurde"
 
 #. Default: "There was an error while updating the reminder"
-#: ./opengever/task/reminder/browser/selectors.py
+#: ./opengever/task/browser/overview.py
 msgid "error_while_updating_task_reminder"
 msgstr "Es ist ein Fehler aufgetreten."
 
@@ -694,7 +694,7 @@ msgid "new_dossier"
 msgstr "neues Dossier"
 
 #. Default: "No reminder"
-#: ./opengever/task/reminder/browser/selectors.py
+#: ./opengever/task/browser/overview.py
 msgid "no_reminder"
 msgstr "Keine Erinnerung"
