+++ opengever/task/locales/fr/LC_MESSAGES/opengever.task.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
-"POT-Creation-Date: 2018-10-04 15:24+0000\n"
+"POT-Creation-Date: 2018-10-05 07:33+0000\n"
 "PO-Revision-Date: 2017-12-03 11:47+0000\n"
 "Last-Translator: Jacqueline Sposato <jacqueline.sposato@gmail.com>\n"
 "Language-Team: French <https://translations.onegovgever.ch/projects/onegov-gever/opengever-task/fr/>\n"
@@ -229,7 +229,7 @@ msgid "error_same_responsible"
 msgstr "Pas de changement: mandataire identique séléctionné"
 
 #. Default: "There was an error while updating the reminder"
-#: ./opengever/task/reminder/browser/selectors.py
+#: ./opengever/task/browser/overview.py
 msgid "error_while_updating_task_reminder"
 msgstr ""
 
@@ -696,7 +696,7 @@ msgid "new_dossier"
 msgstr "Nouveau Dossier"
 
 #. Default: "No reminder"
-#: ./opengever/task/reminder/browser/selectors.py
+#: ./opengever/task/browser/overview.py
 msgid "no_reminder"
 msgstr ""
